+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.0] - 2022-05-09
+
 ### Added
 - Add `frame-pointer` feature to unwind the stack with frame pointer (#116)